+++ src/main/kotlin/ara/analysis/LivenessAnalysis.kt
@@ -8,7 +8,6 @@ import ara.analysis.live.LivenessState.Companion.meet
 import ara.control.Block
 import ara.position.Range
 import ara.reporting.Message.Companion.quoted
-import ara.storage.ResourceAllocation.asResourcePath
 import ara.storage.ResourceAllocation.asResourcePaths
 import ara.storage.ResourceAllocation.resourcesCreated
 import ara.storage.ResourceAllocation.resourcesDestroyed
@@ -45,10 +44,9 @@ class LivenessAnalysis(val program: Syntax.Program) : Analysis<Unit>() {
             for (variable in conflicts.keys.sorted()) {
                 val conflict = conflicts[variable]!!
 
-                // TODO: Check if there is no Conflict in OUT of predecessors, indicating that this is the root cause.
                 val message = reportError("Variable $variable has conflicting initializers and finalizers.")
                 for ((position, definition) in orderedConflictingDefinitions(conflict)) {
-                    message.withAdditionalInfo("A potential cause might be the $definition here:", position)
+                    message.withAdditionalInfo("Potential causes include the $definition here:", position)
                 }
             }
         }
@@ -140,8 +138,7 @@ class LivenessAnalysis(val program: Syntax.Program) : Analysis<Unit>() {
         }
 
         private fun initializeOrReport(expression: Syntax.ResourceExpression) {
-            val resource = expression.asResourcePath()
-            if (resource != null) {
+            for (resource in expression.asResourcePaths()) {
                 val state = currentState[resource]
                 if (state !is LivenessState.Finalized) {
                     val error =
@@ -159,8 +156,7 @@ class LivenessAnalysis(val program: Syntax.Program) : Analysis<Unit>() {
         }
 
         private fun finalizeOrReport(expression: Syntax.ResourceExpression) {
-            val resource = expression.asResourcePath()
-            if (resource != null) {
+            for (resource in expression.asResourcePaths()) {
                 val state = currentState[resource]
                 if (state !is LivenessState.Initialized) {
                     val error =
@@ -191,7 +187,7 @@ class LivenessAnalysis(val program: Syntax.Program) : Analysis<Unit>() {
 
         private fun verifyUses(instruction: Syntax.Instruction) {
             when (instruction) {
-                is Syntax.Assignment -> {
+                is Syntax.ArithmeticAssignment -> {
                     if (instruction.arithmetic != null) {
                         val resources = instruction.arithmetic.value.asResourcePaths().toSet()
                         for (resource in resources) {
@@ -229,6 +225,15 @@ class LivenessAnalysis(val program: Syntax.Program) : Analysis<Unit>() {
             detectedConflicts[variable] = detectedConflicts[variable]!! meet conflict
         }
 
+        private fun variableIsConflictInPredecessors(block: Block, variable: ResourcePath): Boolean {
+            for (predecessor in routine.graph.getPredecessors(block)) {
+                val variableStateAtEndOfPredecessor = routine.liveness.getOut(predecessor)[variable]
+                if (variableStateAtEndOfPredecessor is LivenessState.Conflict)
+                    return true
+            }
+            return false
+        }
+
         private fun detectConflicts() {
             for (block in routine.graph) {
                 val livenessAtEndOfBlock = routine.liveness.getOut(block)
@@ -237,7 +242,9 @@ class LivenessAnalysis(val program: Syntax.Program) : Analysis<Unit>() {
                     val path = ResourcePath.ofIdentifier(variable)
                     val potentialConflict = livenessAtEndOfBlock[path]
 
-                    if (potentialConflict is LivenessState.Conflict) {
+                    if (potentialConflict is LivenessState.Conflict &&
+                        !variableIsConflictInPredecessors(block, path)) {
+                        // Narrow down causes to the first block in which a variable becomes Conflict.
                         updateConflict(variable, potentialConflict)
                     }
                 }