+++ src/main/kotlin/ara/position/Range.kt
@@ -1,12 +1,6 @@
 package ara.position
 
-import kotlin.math.max
-import kotlin.math.min
-
 data class Range(val input: InputSource, val offsetStart: Long, val offsetEnd: Long) : Comparable<Range> {
-
-    infix fun union(range: Range): Range =
-        Range(input, min(this.offsetStart, range.offsetStart), max(this.offsetEnd, range.offsetEnd))
 
     override fun compareTo(other: Range): Int {
         if (this.input.filename() == other.input.filename())
@@ -17,5 +11,4 @@ data class Range(val input: InputSource, val offsetStart: Long, val offsetEnd: L
             other.input.filename()
         )
     }
-
 }