+++ src/main/kotlin/ara/syntax/Typeable.kt
@@ -1,0 +1,5 @@
+package ara.syntax
+
+sealed class Typeable : Syntax() {
+    lateinit var computedType: ara.types.Type
+}