+++ src/main/kotlin/ara/syntax/extensions/Extensions.kt
@@ -1,6 +1,7 @@
 package ara.syntax.extensions
 
 import ara.syntax.Syntax
+import ara.types.Type
 
 
 fun Syntax.RoutineDefinition.isEmpty(): Boolean =
@@ -12,3 +13,8 @@ val Syntax.Program.routines: List<Syntax.RoutineDefinition>
 
 val Syntax.Program.types: List<Syntax.TypeDefinition>
     get() = this.definitions.filterIsInstance<Syntax.TypeDefinition>()
+
+
+fun Syntax.RoutineDefinition.lookupVariableType(variable: Syntax.Identifier): Type? =
+    this.localEnvironment.getVariable(variable)
+