+++ src/main/kotlin/ara/types/Signature.kt
@@ -1,0 +1,3 @@
+package ara.types
+
+data class Signature(val inputTypes: List<Type>, val outputTypes: List<Type>)